+++ index.js
@@ -4,7 +4,6 @@
  *
  * [Proposals]
  * https://babeljs.io/docs/en/babel-plugin-proposal-class-properties
- * https://babeljs.io/docs/en/babel-plugin-proposal-decorators
  * https://babeljs.io/docs/en/babel-plugin-proposal-optional-chaining
  * https://babeljs.io/docs/en/babel-plugin-proposal-nullish-coalescing-operator
  *
@@ -21,6 +20,13 @@
 /**
  * 1. Ensure all helpers are imported instead of inlined.
  *    See https://github.com/babel/babel/issues/9297#issuecomment-453750049
+ *
+ * 2. Plugins are run in the order defined below,
+ *    but they are run *before* presets.
+ *    So we need to make sure the Flow annotations are removed before
+ *    the class properties plugin.
+ *    This is to avoid the `undefined` properties issue with old Flow props
+ *    syntax: `class A { props: Props }`.
  */
 module.exports = (context, options) => {
   let envOpts = Object.assign(
@@ -48,11 +54,10 @@ module.exports = (context, options) => {
   ];
 
   let plugins = [
-    '@babel/plugin-transform-flow-strip-types',
+    '@babel/plugin-transform-flow-strip-types', // 2
+    '@babel/plugin-proposal-class-properties',
     '@babel/plugin-proposal-optional-chaining',
     '@babel/plugin-proposal-nullish-coalescing-operator',
-    ['@babel/plugin-proposal-decorators', { legacy: true }],
-    ['@babel/plugin-proposal-class-properties', { loose: true }],
     ['@babel/plugin-transform-runtime', runtimeOpts],
   ];