+++ test/index.spec.js
@@ -32,13 +32,6 @@ let cases = [
     `,
   ],
   [
-    'decorators legacy',
-    `
-      @decorated
-      class A {}
-    `,
-  ],
-  [
     'dynamic imports',
     `
       import('./a')